renamed team invite form components
